# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/chat/rooms:
    get:
      operationId: listAdminChatRooms
      tags: [admin]
      summary: List language chat rooms and their moderators
      responses:
        "200":
          description: Chat room configuration
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminPublicChatRoom" }
        "401":
          description: Password-authenticated admin session required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      operationId: createAdminChatRoom
      tags: [admin]
      summary: Create a public room for one supported language
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicChatRoomInput" }
      responses:
        "201":
          description: Chat room created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPublicChatRoom" }
        "409":
          description: A room already exists for this language
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /chat/rooms:
    get:
      operationId: listPublicChatRooms
      tags: [chat]
      summary: List active public chat rooms without authentication
      responses:
        "200":
          description: Active language rooms
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublicChatRoom" }
        "503":
          description: Chat room storage is unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /rooms:
    get:
      operationId: listPublicChatRoomsAlias
      tags: [chat]
      summary: List active public chat rooms without authentication (canonical alias)
      responses:
        "200":
          description: Active language rooms
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublicChatRoom" }
        "503":
          description: Chat room storage is unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /chat/messages:
    get:
      operationId: listPublicChatMessages
      tags: [chat]
      summary: List room messages and automatically translate them to the selected app language
      parameters:
        - name: roomId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: targetLanguage
          in: query
          required: true
          schema: { $ref: "#/components/schemas/ChatLanguage" }
        - name: cursor
          in: query
          schema: { type: string, maxLength: 200 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 30 }
      responses:
        "200":
          description: Chronological message page with available translations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicChatMessagePage" }
        "401":
          description: Authentication required
        "404":
          description: Chat room not found or inactive
  /chat/rooms/{roomId}/messages:
    post:
      operationId: sendPublicChatMessage
      tags: [chat]
      summary: Send a message in a public language room
      parameters:
        - $ref: "#/components/parameters/ChatRoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicChatMessageInput" }
      responses:
        "201":
          description: Message stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicChatMessage" }
        "401":
          description: Authentication required
        "404":
          description: Chat room not found or inactive
        "429":
          description: Too many messages; try again later
  /chat/messages/{messageId}:
    delete:
      operationId: moderatePublicChatMessage
      tags: [chat]
      summary: Remove a message as its sender or an assigned room moderator
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Message removed
        "401":
          description: Authentication required
        "403":
          description: Only the message sender, room moderator, or platform admin may remove this message
        "404":
          description: Message not found
...
components:
  securitySchemes:
    piAppSessionCookie:
      type: apiKey
      in: cookie
      name: pi_app_session
  parameters:
...
    ChatRoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ChatLanguage:
      type: string
      enum: [en, ar, zh-CN, id, vi]
    PublicChatRoomInput:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 2, maxLength: 60 }
        languageCode: { $ref: "#/components/schemas/ChatLanguage" }
      required: [name, languageCode]
...
    PublicChatRoom:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        languageCode: { $ref: "#/components/schemas/ChatLanguage" }
        active: { type: boolean }
      required: [id, name, languageCode, active]
    AdminChatModerator:
      type: object
      properties:
        userId: { type: string }
        displayName: { type: string }
        assignedAt: { type: string, format: date-time }
      required: [userId, displayName, assignedAt]
    AdminPublicChatRoom:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        languageCode: { $ref: "#/components/schemas/ChatLanguage" }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        moderators:
          type: array
          items: { $ref: "#/components/schemas/AdminChatModerator" }
      required: [id, name, languageCode, active, createdAt, moderators]
...
    PublicChatMessageInput:
      type: object
      additionalProperties: false
      properties:
        content: { type: string, minLength: 1, maxLength: 2000 }
        targetLanguage: { $ref: "#/components/schemas/ChatLanguage" }
      required: [content, targetLanguage]
    PublicChatMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        roomId: { type: string, format: uuid }
        senderName: { type: string }
        content: { type: ["string", "null"] }
        sourceLanguage: { $ref: "#/components/schemas/ChatLanguage" }
        translatedText: { type: ["string", "null"] }
        translationStatus:
          type: string
          enum: [translated, same_language, unavailable]
        createdAt: { type: string, format: date-time }
        isDeleted: { type: boolean }
        isOwn: { type: boolean }
      required: [id, roomId, senderName, content, sourceLanguage, translatedText, translationStatus, createdAt, isDeleted, isOwn]
    PublicChatMessagePage:
      type: object
      properties:
        messages:
          type: array
          items: { $ref: "#/components/schemas/PublicChatMessage" }
        nextCursor: { type: ["string", "null"] }
        canModerate: { type: boolean }
      required: [messages, nextCursor, canModerate]
...